# api/openapi/openapi.json
{
  "components": {
...
    "schemas": {
...
      "CreateRunRequest": {
        "properties": {
          "category": {
            "description": "Optional flat tag (task t3), e.g. `review`, `audit`, `explore`, `implement`, `docs` — a cheap-to-change label, not a fixed taxonomy, so per-actor aggregates (issue #28) can slice by it. Unlike `name`/`description`, `category` alone is retaggable after creation via PATCH /v1alpha1/runs/{id} (frame decision q4).",
            "type": "string"
          },
          "description": {
            "description": "Optional operator-given free-text description (task t3). Set at creation only, like `name` — immutable afterward.",
            "type": "string"
          },
          "input": {
            "description": "The run input. Validated against the workflow's input contract."
          },
          "name": {
            "description": "Optional operator-given display name (task t3, migrations/0013). Additive: omitting it (every pre-t3 client) behaves exactly as before. Set at creation only — there is no way to change it afterward; see PatchRunRequest.",
            "type": "string"
          },
          "workflow_digest": {
            "type": "string"
          }
        },
        "required": [
          "workflow_digest"
        ],
        "type": "object"
      },
...
      "Error": {
        "description": "The stable error shape, every non-2xx response, every operation. Mirrors internal/clifmt.CliError's JSON encoding exactly: `code` follows the same exit-code-style bucket (1 = domain/user error such as not-found or conflict, 2 = environment error such as a database outage), and the HTTP status line carries the specific condition.",
        "properties": {
          "code": {
            "enum": [
              1,
              2
            ],
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "remediation": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "remediation"
        ],
        "type": "object"
      },
...
      "RunList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Run"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Opaque keyset cursor for the next page; absent on the last page.",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "RunState": {
        "enum": [
          "created",
          "running",
          "waiting",
          "completed",
          "failed",
          "cancelled"
        ],
        "type": "string"
      },
...
      "Ticket": {
        "properties": {
          "frozen": {
            "type": "boolean"
          },
          "human_tasks": {
            "description": "Every human task on the ticket's runs, in any status.",
            "items": {
              "$ref": "#/components/schemas/HumanTask"
            },
            "type": "array"
          },
          "latest_frame": {
            "$ref": "#/components/schemas/TicketFrame"
          },
          "ledger": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "merged_pr": {
            "additionalProperties": true,
            "type": "object"
          },
          "pending_tasks": {
            "description": "The decidable subset of human_tasks. Both are served because they answer different questions — what was asked, and what is still being asked.",
            "items": {
              "$ref": "#/components/schemas/TicketPendingTask"
            },
            "type": "array"
          },
          "replies": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/Run"
            },
            "type": "array"
          },
          "ticket_id": {
            "type": "string"
          },
          "ticket_reports": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "ticket_url": {
            "description": "Where this ticket lives on the board, composed from the Jira work-item fact the ticket's runs carry (`details_url`, else `jira_site` plus the key), falling back to a posted frame's own `ticket_url`/`jira_url` (task t18, spec c10). Always an absolute http(s) URL; absent when nothing this projection can see says where the ticket lives.",
            "type": "string"
          }
        },
        "required": [
          "ticket_id",
          "runs",
          "ledger",
          "human_tasks",
          "pending_tasks",
          "ticket_reports",
          "replies",
          "frozen"
        ],
        "type": "object"
      },
      "TicketFrame": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "frame": {
            "additionalProperties": true,
            "type": "object"
          },
          "posted_by": {
            "type": "string"
          },
          "ticket_id": {
            "type": "string"
          },
          "version": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "ticket_id",
          "version",
          "frame",
          "posted_by",
          "created_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "actorRegistrationBearerAuth": {
        "description": "The shared secret configured as NODES_ACTOR_REGISTRATION_TOKEN_SECRET. Used only by POST /v1alpha1/actors (task t13) — deliberately its own secret, not the decision secret, so an operator can grant registration standing without also granting the power to decide human tasks.",
        "scheme": "bearer",
        "type": "http"
      },
      "adhocRunBearerAuth": {
        "description": "The shared secret configured as NODES_ADHOC_RUN_TOKEN_SECRET. Used only by POST /v1alpha1/adhoc-runs (task t19, gated by the t15 auth-hardening pass, spec c27) — its own secret, so the power to start ad-hoc (often billable) work is granted independently of decision, registration, and event standing.",
        "scheme": "bearer",
        "type": "http"
      },
      "attemptCallbackBearerAuth": {
        "description": "The short-lived callback token minted for exactly one attempt. The route verifies it against the attempt id in the path before loading durable invocation authority or reading the request body.",
        "scheme": "bearer",
        "type": "http"
      },
      "decisionBearerAuth": {
        "description": "The shared secret configured as NODES_HUMAN_DECISION_TOKEN_SECRET. Used by the three operations that record a human decision — POST /v1alpha1/human-tasks/{id}/decision and the two review operations (POST /v1alpha1/runs/{id}/reviews, POST /v1alpha1/reviews/{id}/commit) — because all three write human-authority records into the ledger on whoever presents the token. Every other operation in this API is authless by phase-1 design (spec decision c45).",
        "scheme": "bearer",
        "type": "http"
      },
      "eventBearerAuth": {
        "description": "The shared secret configured as NODES_EVENT_TOKEN_SECRET. Used only by POST /v1alpha1/events (task t10) — its own secret again, so an external system granted the standing to emit signal events (which can resume any signal-parked run) holds neither decision nor registration power.",
        "scheme": "bearer",
        "type": "http"
      },
      "inboundActorKeyAuth": {
        "description": "Names the party whose credential is being presented — an actor key (`namespace/name`), which is also the mailbox partition the poll claims from and the completion is scoped to. Required together with inboundDialBearerAuth: an absent or empty header is a 401 before the credential is examined at all. The transport routes hardcode the party KIND to `actor`, so a `host`-kind credential cannot dial in even though issuance will mint one.",
        "in": "header",
        "name": "X-Culture-Nodes-Actor-Key",
        "type": "apiKey"
      },
      "inboundDialBearerAuth": {
        "description": "The credential the control plane ISSUED for one dialling party — the `cnd_`-prefixed value returned exactly once by POST /v1alpha1/inbound/credentials. Presented by a bridge on every poll and every completion, alongside inboundActorKeyAuth; neither header is sufficient alone. Verified against a SHA-256 verifier read from PostgreSQL per call, never cached, which is what makes revocation effective on the next dial.",
        "scheme": "bearer",
        "type": "http"
      },
      "inboundIssuanceBearerAuth": {
        "description": "The shared secret configured as NODES_INBOUND_ISSUANCE_TOKEN_SECRET. Used only by POST /v1alpha1/inbound/credentials and POST /v1alpha1/inbound/credentials/revoke (issue #111's dial-in half) — its own secret again, so the standing to mint or end a bridge's IDENTITY is granted independently of decision, registration, event and ad-hoc power. Absent configuration leaves both routes refusing with 401 rather than mounted-but-authless.",
        "scheme": "bearer",
        "type": "http"
...
  },
  "info": {
    "description": "The Culture Nodes control-plane API (docs/initial-design/culture-nodes-prd-spec.md §12.1, §15.1): workflow publication, run orchestration, the append-only work ledger, and review transactions. group: nodes.culture.dev/v1alpha1.\nPhase 1 runs authless by design (spec decision c45): this API carries no authentication or authorization of its own and is meant to run behind a private network. Every operation is reachable by anyone who can reach the listener. Do not expose this API on a public network without a fronting proxy that adds authentication.\nEvery error response uses the Error schema below — the same {code, message, remediation} shape internal/clifmt.CliError renders for the CLI — with the HTTP status line carrying the specific condition. The two reverse-transport routes are the documented exception: POST /v1alpha1/inbound/poll and POST /v1alpha1/inbound/{id}/complete are not mounted through the wrapper that renders Error, so their 400/401 refusals are `text/plain` while their 500s still carry Error. Each of those responses says so where it is declared.",
    "license": {
      "name": "Apache-2.0"
    },
    "summary": "Culture Nodes control-plane API",
    "title": "nodes.culture.dev/v1alpha1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1alpha1/runs": {
      "get": {
        "description": "Newest first by `sort` (task t11). `updated_since`/`updated_until` filter on `updated_at`, using `runs_namespace_updated_at_idx` (migrations/0010) when `sort` is (or defaults to) `updated_at`.",
        "operationId": "listRuns",
        "parameters": [
          {
            "description": "Filter to runs currently in this state.",
            "in": "query",
            "name": "state",
            "schema": {
              "$ref": "#/components/schemas/RunState"
            }
          },
          {
            "description": "Filter to runs pinned to a workflow with this exact key.",
            "in": "query",
            "name": "workflow_key",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter to runs whose persisted correlation subject matches exactly.",
            "in": "query",
            "name": "subject",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only runs updated at or after this instant (RFC3339). When set and `sort` is omitted, `sort` defaults to `updated_at` rather than the endpoint's own default of `created_at`.",
            "in": "query",
            "name": "updated_since",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Only runs updated at or before this instant (RFC3339). When set and `sort` is omitted, `sort` defaults to `updated_at` rather than the endpoint's own default of `created_at`.",
            "in": "query",
            "name": "updated_until",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Sort column: `created_at` (the endpoint's own default) or `updated_at`. Always descending — newest first, like every list in this API. Defaults to `updated_at` instead when either `updated_since` or `updated_until` is set and `sort` itself is omitted.",
            "in": "query",
            "name": "sort",
            "schema": {
              "enum": [
                "created_at",
                "updated_at"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque keyset cursor from a previous response's next_cursor.",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            },
            "description": "A page of runs, newest first."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "state, updated_since, updated_until, sort, or cursor is malformed."
          }
        },
        "summary": "List runs, optionally filtered by workflow key, state, and/or an updated_at window.",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Pins the run to the given workflow digest's immutable normalized IR (PRD §20.1) and creates the entry token and node run in one transaction (PRD §12.5). The run is claimable the moment this returns.",
        "operationId": "createRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "The run was created and is now running."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The input does not satisfy the workflow's input contract."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No workflow version has the given digest."
          }
        },
        "summary": "Create a run of a published workflow.",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1alpha1/tickets/{id}": {
      "get": {
        "operationId": "getTicket",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            },
            "description": "Runs with this subject, their ledger and human tasks, ticket reports, replies, and the latest frame."
          }
        },
        "summary": "Compose the control-plane projection for one ticket.",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "The API is served relative to whatever host and port `nodes serve` binds.",
      "url": "/"
    }
  ],
...
}